Implemented Access Control

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use DB;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -76,6 +86,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // Check for correct user
+        if (auth()->user()->id !== $post->user_id):
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        endif;
         return view('posts.edit')->with('post', $post);
     }
 
@@ -109,6 +123,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if (auth()->user()->id !== $post->user_id):
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        endif;
+        
         $post->delete();
         return redirect('./posts')->with('error', 'Post Removed!');
     }
